Workaround for DAFFODIL-2974
This schema no longer defines a root element for use by other schemas.

Rather, it defines the root element, just for testing, in no namespace under src/test only.

To use this schema, you use the complexType messageType in an element with name of your own choosing.

This was required due to bug DAFFODIL-2947